pyproject.toml
[project]
name = "llmd-python"
version = "0.1.0"
description = "Python SDK for LLMD (LLM Dialogue) format - create, parse, and manipulate conversation files"
readme = "README.md"
authors = [
    { name = "LLMD Format Team", email = "team@llmd-format.org" }
]
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
    "typing-extensions>=4.0.0; python_version<'3.11'",
]
keywords = ["llm", "dialogue", "conversation", "format", "ai", "chatbot"]
license = { text = "Apache-2.0" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Markup",
]

[project.urls]
Homepage = "https://github.com/llmd-format/sdk-python"
Repository = "https://github.com/llmd-format/sdk-python"
Documentation = "https://github.com/llmd-format/sdk-python#readme"
"Bug Tracker" = "https://github.com/llmd-format/sdk-python/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "black>=23.0",
    "isort>=5.0",
    "flake8>=6.0",
]

[project.scripts]
llmd = "llmd_python.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=llmd_python",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["llmd_python"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
]
